requirements.txt
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
